refactor: move generic
