+++ crates/prql-compiler/src/ir/mod.rs
@@ -1,7 +1,7 @@
 //! Intermediate Representations of Abstract Syntax Tree
 //!
+pub mod generic;
 pub mod pl;
 pub mod rq;
 
-pub use prql_ast::expr::generic;
 pub use prql_ast::Span;
