+++ crates/prql-compiler/src/sql/gen_expr.rs
@@ -9,13 +9,13 @@ use sqlparser::ast::{
 };
 use std::cmp::Ordering;
 
-use crate::generic::{ColumnSort, SortDirection, WindowFrame, WindowKind};
-use crate::ir::generic::{InterpolateItem, Range};
+use crate::ir::generic::{ColumnSort, SortDirection, WindowFrame, WindowKind};
 use crate::ir::pl::{self, Ident, Literal};
 use crate::ir::rq::*;
 use crate::sql::srq::context::ColumnDecl;
 use crate::utils::{OrMap, VALID_IDENT};
 use crate::{Error, Span, WithErrorInfo};
+use prql_ast::expr::generic::{InterpolateItem, Range};
 
 use super::gen_projection::try_into_exprs;
 use super::{keywords, Context};
@@ -933,8 +933,8 @@ impl From<sql_ast::Expr> for ExprOrSource {
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::ir::generic::Range;
     use insta::assert_yaml_snapshot;
+    use prql_ast::expr::generic::Range;
 
     #[test]
     fn test_range_of_ranges() -> Result<()> {
