+++ crates/prql-compiler/src/sql/srq/preprocess.rs
@@ -5,8 +5,7 @@ use std::collections::HashSet;
 use anyhow::Result;
 use itertools::Itertools;
 
-use crate::generic::{ColumnSort, SortDirection, WindowFrame, WindowKind};
-use crate::ir::generic::{InterpolateItem, Range};
+use crate::ir::generic::{ColumnSort, SortDirection, WindowFrame, WindowKind};
 use crate::ir::pl::{JoinSide, Literal};
 use crate::ir::rq::{
     self, maybe_binop, new_binop, CId, Compute, Expr, ExprKind, RqFold, TableRef, Transform, Window,
@@ -14,6 +13,7 @@ use crate::ir::rq::{
 use crate::sql::srq::context::ColumnDecl;
 use crate::sql::Context;
 use crate::{Error, WithErrorInfo};
+use prql_ast::expr::generic::{InterpolateItem, Range};
 
 use super::anchor::{infer_complexity, CidCollector, Complexity};
 use super::ast::*;
